+++ runtime/src/comm/none/comm-none.c
@@ -143,11 +143,11 @@ void chpl_stopCommDiagnosticsHere() { }
 void chpl_resetCommDiagnosticsHere() { }
 void chpl_getCommDiagnosticsHere(chpl_commDiagnostics *cd) { }
 
-int32_t chpl_numCommGets(void) { return 0; }
-int32_t chpl_numCommNBGets(void) { return 0; }
-int32_t chpl_numCommTestNBGets(void) { return 0; }
-int32_t chpl_numCommWaitNBGets(void) { return 0; }
-int32_t chpl_numCommPuts(void) { return 0; }
-int32_t chpl_numCommForks(void) { return 0; }
-int32_t chpl_numCommFastForks(void) { return 0; }
-int32_t chpl_numCommNBForks(void) { return 0; }
+uint64_t chpl_numCommGets(void) { return 0; }
+uint64_t chpl_numCommNBGets(void) { return 0; }
+uint64_t chpl_numCommTestNBGets(void) { return 0; }
+uint64_t chpl_numCommWaitNBGets(void) { return 0; }
+uint64_t chpl_numCommPuts(void) { return 0; }
+uint64_t chpl_numCommForks(void) { return 0; }
+uint64_t chpl_numCommFastForks(void) { return 0; }
+uint64_t chpl_numCommNBForks(void) { return 0; }
